Add XmlNodeName attribute
TimeLog.ReportingApi.Exporter compares XML node name against the property name in TimeLog.ReportingApi.SDK to get the right value for the export columns.
We have some columns implemented with different property name from the XML. So XmlNodeName attribute is created.
This attribute is used when the XML node name is different from the property name, to facilitate the export name matching.
Due to backwards compatibility issue, renaming the property is not a valid solution